requirements.txt
pyautogen
openai
matplotlib
wordcloud
streamlit